refactor(tui): move startup utils out of constant/kimi-tui
combineStartupNotice 和 isOAuthLoginRequiredError 是工具函数不是常量，
从 constant/kimi-tui.ts 移到 utils/startup.ts。
